moved SelectionState to new file. It now is also responsible for logic connected to selections.
