+++ app/src/main/java/com/ethran/notable/utils/EditorState.kt
@@ -1,21 +1,13 @@
 package com.ethran.notable.utils
 
-import android.graphics.Bitmap
-import android.graphics.Canvas
 import android.graphics.Color
-import android.graphics.Rect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.unit.IntOffset
-import androidx.core.graphics.createBitmap
 import com.ethran.notable.classes.ClipboardContent
 import com.ethran.notable.classes.PageView
-import com.ethran.notable.classes.showHint
+import com.ethran.notable.classes.SelectionState
 import com.ethran.notable.datastore.EditorSettingCacheManager
-import com.ethran.notable.db.Image
-import com.ethran.notable.db.Stroke
-import kotlinx.coroutines.CoroutineScope
 
 enum class Mode {
     Draw, Erase, Select, Line
@@ -63,82 +55,6 @@ class EditorState(val bookId: String? = null, val pageId: String, val pageView: 
 enum class PlacementMode {
     Move,
     Paste
-}
-
-class SelectionState {
-    var firstPageCut by mutableStateOf<List<SimplePointF>?>(null)
-    var secondPageCut by mutableStateOf<List<SimplePointF>?>(null)
-    var selectedStrokes by mutableStateOf<List<Stroke>?>(null)
-    var selectedImages by mutableStateOf<List<Image>?>(null)
-    var selectedBitmap by mutableStateOf<Bitmap?>(null)
-    var selectionStartOffset by mutableStateOf<IntOffset?>(null)
-    var selectionDisplaceOffset by mutableStateOf<IntOffset?>(null)
-    var selectionRect by mutableStateOf<Rect?>(null)
-    var placementMode by mutableStateOf<PlacementMode?>(null)
-
-    fun reset() {
-        selectedStrokes = null
-        selectedImages = null
-        secondPageCut = null
-        firstPageCut = null
-        selectedBitmap = null
-        selectionStartOffset = null
-        selectionRect = null
-        selectionDisplaceOffset = null
-        placementMode = null
-    }
-
-    fun isResizable(): Boolean {
-        return selectedImages?.count() == 1 && selectedStrokes.isNullOrEmpty()
-    }
-
-    fun resizeImages(scale: Int, scope: CoroutineScope, page: PageView){
-        val selectedImagesCopy = selectedImages?.map { image ->
-            image.copy(
-                height = image.height + (image.height * scale / 100),
-                width = image.width + (image.width * scale / 100)
-            )
-        }
-
-        // Ensure selected images are not null or empty
-        if (selectedImagesCopy.isNullOrEmpty()) {
-            showHint("For now, strokes cannot be resized", scope)
-            return
-        }
-
-        selectedImages = selectedImagesCopy
-        // Adjust displacement offset by half the size change
-        val sizeChange = selectedImagesCopy.firstOrNull()?.let { image ->
-            IntOffset(
-                x = (image.width * scale / 200),
-                y = (image.height * scale / 200)
-            )
-        } ?: IntOffset.Zero
-
-        val pageBounds = imageBoundsInt(selectedImagesCopy)
-        selectionRect = pageAreaToCanvasArea(pageBounds, page.scroll)
-
-        selectionDisplaceOffset =
-            selectionDisplaceOffset?.let { it - sizeChange }
-                ?: IntOffset.Zero
-
-        val selectedBitmapNew = createBitmap(pageBounds.width(), pageBounds.height())
-        val selectedCanvas = Canvas(selectedBitmapNew)
-        selectedImagesCopy.forEach {
-            drawImage(
-                page.context,
-                selectedCanvas,
-                it,
-                IntOffset(-it.x, -it.y)
-            )
-        }
-
-        // set state
-        selectedBitmap = selectedBitmapNew
-    }
-    fun resizeStrokes(scale: Int, scope: CoroutineScope, page: PageView){
-        //TODO: implement this
-    }
 }
 
 object Clipboard {
